add portable mode

+++ src/main.rs
@@ -54,18 +54,54 @@ struct Args {
 }
 
 fn main() {
-    let config_dir = dirs::config_dir().unwrap().join("gopher64");
-    let cache_dir = dirs::cache_dir().unwrap().join("gopher64");
+    let portable = dirs::executable_dir()
+        .unwrap()
+        .join("portable.txt")
+        .exists();
+    let config_dir;
+    let cache_dir;
+    let data_dir;
+    if portable {
+        config_dir = dirs::executable_dir()
+            .unwrap()
+            .join("portable_data")
+            .join("config");
+        cache_dir = dirs::executable_dir()
+            .unwrap()
+            .join("portable_data")
+            .join("cache");
+        data_dir = dirs::executable_dir()
+            .unwrap()
+            .join("portable_data")
+            .join("data");
+    } else {
+        config_dir = dirs::config_dir().unwrap().join("gopher64");
+        cache_dir = dirs::cache_dir().unwrap().join("gopher64");
+        data_dir = dirs::data_dir().unwrap().join("gopher64");
+    };
 
-    let _ = std::fs::create_dir_all(config_dir.clone());
-    let _ = std::fs::create_dir_all(cache_dir.clone());
-    let _ = std::fs::remove_file(cache_dir.clone().join("game_running"));
+    let mut result = std::fs::create_dir_all(config_dir.clone());
+    if result.is_err() {
+        panic!("could not create config dir")
+    }
+    result = std::fs::create_dir_all(cache_dir.clone());
+    if result.is_err() {
+        panic!("could not create cache dir")
+    }
+    result = std::fs::create_dir_all(data_dir.clone());
+    if result.is_err() {
+        panic!("could not create data dir")
+    }
+    result = std::fs::remove_file(cache_dir.clone().join("game_running"));
+    if result.is_err() {
+        panic!("could not remove running file")
+    }
 
     let args = Args::parse();
     let args_as_strings: Vec<String> = std::env::args().collect();
     let args_count = args_as_strings.len();
     if args_count > 1 {
-        let mut device = device::Device::new();
+        let mut device = device::Device::new(config_dir);
 
         if args.clear_input_bindings {
             ui::input::clear_bindings(&mut device.ui);
@@ -116,7 +152,7 @@ fn main() {
 
         if args.game.is_some() {
             let file_path = std::path::Path::new(args.game.as_ref().unwrap());
-            device::run_game(file_path, &mut device, args.fullscreen);
+            device::run_game(file_path, data_dir, &mut device, args.fullscreen);
         }
     } else {
         let options = eframe::NativeOptions {
@@ -126,7 +162,11 @@ fn main() {
         eframe::run_native(
             "gopher64",
             options,
-            Box::new(|cc| Ok(Box::new(ui::gui::GopherEguiApp::new(cc)))),
+            Box::new(|cc| {
+                Ok(Box::new(ui::gui::GopherEguiApp::new(
+                    cc, config_dir, cache_dir, data_dir,
+                )))
+            }),
         )
         .unwrap();
     }
